requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
circuitbreaker==2.0.0
Django==5.1.2
django-db-connection-pool==1.2.5
django-db-geventpool==4.0.7
django-tenants==3.7.0
greenlet==3.1.1
idna==3.10
mysql-connector-python==9.1.0
postgres==4.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pybreaker==1.2.0
requests==2.32.3
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
